requirements.txt
certifi
charset-normalizer
dnspython
idna
ipfs-api
numpy
paho-mqtt
pymongo
python-dotenv
requests
schedule
scipy
six
urllib3
